Break out Bot class, fix client
